// src copy/pages/Frontpage/types.tsx
import { SystemStyleObject } from "@styled-system/css";

export type FrontPageProps = {
    styleConfig: {
        titleBox: SystemStyleObject;
        pageTitle: SystemStyleObject;
        mainContentContainer: SystemStyleObject;
        loanInfoContainer: SystemStyleObject;
        newsContainer: SystemStyleObject;
        creditRaiseContainer: SystemStyleObject;
        nostoContainer: SystemStyleObject;
        nostoImage: SystemStyleObject;
        nostoText: SystemStyleObject;
    };
};